@ChoMinGi refactor: Move S3s, Images, Reviews, Tags from feature_review to each package
