+++ matgpt/src/main/java/com/ktc/matgpt/chatgpt/service/GptService.java
@@ -9,8 +9,8 @@ import com.ktc.matgpt.chatgpt.entity.GptReview;
 import com.ktc.matgpt.chatgpt.exception.ApiException;
 import com.ktc.matgpt.chatgpt.repository.GptGuidanceRepository;
 import com.ktc.matgpt.chatgpt.repository.GptReviewRepository;
-import com.ktc.matgpt.feature_review.review.ReviewService;
-import com.ktc.matgpt.feature_review.review.entity.Review;
+import com.ktc.matgpt.review.ReviewService;
+import com.ktc.matgpt.review.entity.Review;
 import com.ktc.matgpt.store.StoreService;
 import com.ktc.matgpt.user.service.UserService;
 import lombok.RequiredArgsConstructor;
