--- matgpt/src/main/java/com/ktc/matgpt/feature_review/s3/S3Service.java
@@ -1,35 +1,0 @@
-package com.ktc.matgpt.feature_review.s3;
-
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.ObjectMetadata;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
-import java.io.IOException;
-
-@RequiredArgsConstructor
-@Service
-public class S3Service {
-    private final AmazonS3 amazonS3;
-
-    @Value("${cloud.aws.s3.bucket}")
-    private String bucket;
-
-    public String save(MultipartFile multipartFile) throws IOException {
-        String filename = multipartFile.getOriginalFilename();
-
-        ObjectMetadata metadata = new ObjectMetadata();
-        metadata.setContentLength(multipartFile.getSize());
-        metadata.setContentType(multipartFile.getContentType());
-        
-        amazonS3.putObject(bucket, filename, multipartFile.getInputStream(), metadata);
-
-        return amazonS3.getUrl(bucket, filename).toString();
-    }
-
-    // TODO: s3 삭제 구현
-    public void deleteImage(String originalFilename)  {
-        amazonS3.deleteObject(bucket, originalFilename);
-    }
-}
