+++ matgpt/src/main/java/com/ktc/matgpt/review/ReviewJPARepository.java
@@ -1,15 +1,13 @@
-package com.ktc.matgpt.feature_review.review;
+package com.ktc.matgpt.review;
 
 
-import com.ktc.matgpt.feature_review.review.entity.Review;
+import com.ktc.matgpt.review.entity.Review;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
