+++ matgpt/src/main/java/com/ktc/matgpt/review/entity/PriceRange.java
@@ -1,4 +1,4 @@
-package com.ktc.matgpt.feature_review.review.entity;
+package com.ktc.matgpt.review.entity;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
