+++ matgpt/src/main/java/com/ktc/matgpt/review/entity/Review.java
@@ -1,4 +1,4 @@
-package com.ktc.matgpt.feature_review.review.entity;
+package com.ktc.matgpt.review.entity;
 
 import com.ktc.matgpt.store.Store;
 import com.ktc.matgpt.utils.BaseTimeEntity;
